requirements.txt
fastapi==0.111.0
uvicorn==0.35.0
paddleocr==3.1.0
paddlepaddle>=2.6.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
bcrypt==4.1.2
#pydantic==1.10.14
#pydantic-settings==1.0.1